Сделал два компонента-контейнера
